+++ src/components/ContextProvider.tsx
@@ -1,10 +1,10 @@
-import React from "react";
+import React, { useRef } from "react";
 import { createContext, useContext, useState } from "react";
 const Context = createContext<any>({});
-/*
- * allows components to access and change variables "history" and "width"
- * width is used to rerender the app to fit device size
- * history is used for the thread arrow
+/**
+ * Holds global application values.
+ * @param props - { children: JSX.Element }
+ * @returns The ContextProvider is returning a JSX element.
  */
 export default function ContextProvider(props: { children: JSX.Element }) {
   const [history, setHistory] = useState("");
@@ -12,11 +12,15 @@ export default function ContextProvider(props: { children: JSX.Element }) {
   const [width, setWidth] = useState(window.innerWidth);
   const [height, setHeight] = useState(window.innerHeight);
   const [notification, setNotification] = useState({ open: false, text: "" });
+  const resizehandler = useRef(false);
   function updateSize() {
     setWidth(window.innerWidth);
     setHeight(window.innerHeight);
   }
-  window.addEventListener("resize", updateSize);
+  if (!resizehandler.current) {
+    resizehandler.current = true;
+    window.addEventListener("resize", updateSize);
+  }
   return (
     <Context.Provider
       value={{
@@ -31,6 +35,10 @@ export default function ContextProvider(props: { children: JSX.Element }) {
     </Context.Provider>
   );
 }
+/**
+ * It returns the current history and a setter for the history
+ * @returns The history object and a setter function.
+ */
 export function useHistory(): [
   string,
   React.Dispatch<React.SetStateAction<string>>
@@ -38,6 +46,18 @@ export function useHistory(): [
   const { history } = useContext(Context);
   return history;
 }
+/**
+ * "Use the width of the window."
+ *
+ * The function returns a tuple of two values:
+ *
+ * The width of the window
+ * A setter function for the width of the window
+ *
+ * The first value is the width of the window. The second value is a setter function for the width of
+ * the window
+ * @returns The width of the window and a setter for the width of the window.
+ */
 export function useWidth(): [
   number,
   React.Dispatch<React.SetStateAction<number>>
@@ -45,6 +65,11 @@ export function useWidth(): [
   const { width } = useContext(Context);
   return width;
 }
+/**
+ * It returns a tuple of the last search query and a setter function for the query
+ * @returns A tuple of two values. The first value is the query, and the second value is a
+ * setter function for the query.
+ */
 export function useQuery(): [
   string,
   React.Dispatch<React.SetStateAction<string>>
@@ -52,6 +77,15 @@ export function useQuery(): [
   const { query } = useContext(Context);
   return query;
 }
+/**
+ * "Use the height value from the Context."
+ *
+ * The function returns a tuple of two values:
+ *
+ * The height value from the Context.
+ * A setter function for the height value in the Context
+ * @returns The height of the window and a setter for the height of the window.
+ */
 export function useHeight(): [
   number,
   React.Dispatch<React.SetStateAction<number>>
@@ -59,6 +93,12 @@ export function useHeight(): [
   const { height } = useContext(Context);
   return height;
 }
+/**
+ * It returns a tuple of two values. The first value is an object that represents the state of the
+ * notification. The second value is a function that sets the state of the notification
+ * @returns A tuple of two values. The first value is the current state of the notification. The second
+ * value is a function that can be used to update the state of the notification.
+ */
 export function useNotification(): [
   { open: boolean; text: string },
   React.Dispatch<React.SetStateAction<{ open: boolean; text: string }>>