+++ src/components/conversation.tsx
@@ -1,5 +1,13 @@
 import "./css/conversation.css";
-import React, { memo, useEffect, useRef, useState } from "react";
+import React, {
+  createContext,
+  memo,
+  SetStateAction,
+  useContext,
+  useEffect,
+  useRef,
+  useState,
+} from "react";
 import {
   Box,
   Button,
@@ -9,7 +17,7 @@ import {
   SelectChangeEvent,
 } from "@mui/material";
 import queryString from "query-string";
-import Comment from "./comment";
+import Comment from "./conversation/comment";
 import Title from "./conversation/title";
 import axios, { AxiosError } from "axios";
 import DOMPurify from "dompurify";
@@ -21,36 +29,60 @@ import { useCat, useId } from "./MenuProvider";
 import { useNotification } from "./ContextProvider";
 import Share from "./conversation/share";
 import { ShareProvider } from "./ShareProvider";
+import PageBottom from "./conversation/pagebottom";
+const ConversationContext = createContext<any>(null);
 type comment = {
+  /** comment id */
   id: number;
+  /** comment user id */
   user: number;
+  /** the comment (in stringified html) */
   comment: string;
-  createdAt: string; //date string
+  /** date string */
+  createdAt: string;
+  /** number of downvotes */
   D?: number;
+  /** number of upvotes */
   U?: number;
+};
+type details = {
+  /** thread shortened link */
+  slink?: string;
+  /** thread category id */
+  category?: number;
+  /** thread title */
+  title?: string;
+  /** number of comments */
+  c?: number;
+  /** thread original poster name */
+  op?: string;
 };
 /**
  * Gets data from /api/thread/<thread id(props.id)>/<conversation/users>
  * Then renders it as Comments
- * @param props.id the thread id
+ * @param {number} props.id the thread id
  * @returns full conversation as Comments
  */
 function Conversation(props: { id: number }) {
   const query = queryString.parse(window.location.search);
   const [notification, setNotification] = useNotification();
   const [conversation, setConversation] = useState<comment[]>([]);
-  const [page, setPage] = useState(Number(query.page) || 1);
+  const [page, setPage] = useState(
+    Number(query.page) || Math.floor(Number(query.c) / 25) + 1 || 1
+  );
   const [users, setUsers] = useState<any>({});
-  const [details, setDetails] = useState<any>({});
+  const [details, setDetails] = useState<details>({});
   const [votes, setVotes] = useState<any>({});
   const [updating, setUpdating] = useState(false);
   const [pages, setPages] = useState(1);
   const [end, setEnd] = useState(false);
   const [n, setN] = useState(Math.random());
+  const [story, setStory] = useState(0);
   const lastHeight = useRef(0);
   const [cat, setCat] = useCat();
   const [id, setId] = useId();
   const navigate = useNavigate();
+  /* Checking if the error is a 404 error and if it is, it will navigate to the 404 page. */
   const onError = function (err: AxiosError) {
     !notification.open &&
       setNotification({
@@ -60,15 +92,40 @@ function Conversation(props: { id: number }) {
     err?.response?.status === 404 && navigate("/404", { replace: true });
   };
   !query.page &&
+    !query.c &&
     navigate(`${window.location.pathname}?page=1`, { replace: true });
   useEffect(() => {
     axios
       .get(`/api/thread/${props.id}?type=1`)
       .then((res) => {
-        setDetails(res.data);
+        res.data.slink && setDetails(res.data);
         !cat && setCat(res.data.category);
         id !== res.data.id && setId(res.data.id);
         document.title = `${res.data.title} | Metahkg`;
+        if (!res.data.slink) {
+          axios
+            .post("https://api-us.wcyat.me/create", {
+              url: `${window.location.origin}/thread/${props.id}?page=1`,
+            })
+            .then((sres) => {
+              setDetails(
+                Object.assign(res.data, {
+                  slink: `https://l.wcyat.me/${sres.data.id}`,
+                })
+              );
+            })
+            .catch(() => {
+              setNotification({
+                open: true,
+                text: "Unable to generate shortened link. A long link will be used instead.",
+              });
+              setDetails(
+                Object.assign(res.data, {
+                  slink: `${window.location.origin}/thread/${props.id}?page=1`,
+                })
+              );
+            });
+        }
       })
       .catch(onError);
     axios
@@ -80,14 +137,15 @@ function Conversation(props: { id: number }) {
     axios
       .get(`/api/thread/${props.id}?type=2&page=${page}`)
       .then((res) => {
-        res.data?.[0] === null && navigate("/404", {replace: true});
+        /** redirect to 404 if thread (or page) not found */
+        res.data?.[0] === null && navigate("/404", { replace: true });
         setConversation(res.data);
         res.data.length % 25 && setEnd(true);
       })
       .catch(onError);
     if (localStorage.user) {
       axios
-        .get(`/api/getvotes?id=${Number(props.id)}`)
+        .get(`/api/getvotes?id=${props.id}`)
         .then((res) => {
           setVotes(res.data);
         })
@@ -96,7 +154,8 @@ function Conversation(props: { id: number }) {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [n]);
   /**
-   * @description get new comments or next page (if last comment id % 25 = 0)
+   * It fetches new comments, or the next page (if last comment id % 25 = 0)
+   * of messages from the server and appends them to the conversation array
    */
   function update() {
     setUpdating(true);
@@ -132,24 +191,13 @@ function Conversation(props: { id: number }) {
           setPage((page) => page + 1);
           setPages(Math.floor((conversation.length - 1) / 25) + 1);
           navigate(`/thread/${props.id}?page=${page + 1}`, { replace: true });
-          document.getElementById(String(page + 1))?.scrollIntoView();
-          const croot = document.getElementById("croot");
-          if (
-            // @ts-ignore
-            !(croot?.clientHeight / 5 + 60 >= croot?.clientHeight) &&
-            // @ts-ignore
-            croot?.scrollHeight - croot?.scrollTop > croot?.clientHeight
-          ) {
-            // @ts-ignore
-            croot.scrollTop -= croot?.clientHeight / 5;
-          }
         }
         setUpdating(false);
       });
   }
   /**
    * It takes a page number and sends a request to the server to get the next page of comments
-   * @param {number} p - number
+   * @param {number} p - new page number
    */
   function changePage(p: number) {
     setConversation([]);
@@ -185,13 +233,21 @@ function Conversation(props: { id: number }) {
       }
     }
   }
+  /* It's checking if the conversation, users, details and votes are all ready. */
   const ready = !!(
     conversation.length &&
     Object.keys(users).length &&
     Object.keys(details).length &&
     (localStorage.user ? Object.keys(votes).length : 1)
   );
-
+  if (ready && query.c) {
+    navigate(`${window.location.pathname}?page=${page}`, {
+      replace: true,
+    });
+    setTimeout(() => {
+      document.getElementById(`c${query.c}`)?.scrollIntoView();
+    }, 100);
+  }
   return (
     <ShareProvider>
       <div className="min-height-fullvh conversation-root">
@@ -205,10 +261,11 @@ function Conversation(props: { id: number }) {
         <Paper
           id="croot"
           key={n}
-          className="overflow-auto conversation-paper"
+          className="overflow-auto nobgimage noshadow conversation-paper"
+          sx={{ bgcolor: "primary.dark" }}
           onScroll={onScroll}
         >
-          <Box className="fullwidth" sx={{ bgcolor: "primary.dark" }}>
+          <Box className="fullwidth max-height-full max-width-full">
             {ready &&
               [...Array(pages)].map((p, index) => (
                 <Box>
@@ -248,7 +305,7 @@ function Conversation(props: { id: number }) {
                   >
                     <PageTop
                       id={roundup(conversation[0].id / 25) + index}
-                      pages={roundup(details.c / 25)}
+                      pages={roundup((details.c || 0) / 25)}
                       page={roundup(conversation[0].id / 25) + index}
                       onChange={(e: SelectChangeEvent<number>) => {
                         changePage(Number(e.target.value));
@@ -261,7 +318,7 @@ function Conversation(props: { id: number }) {
                       }
                       next={
                         roundup(conversation[0].id / 25) + index !==
-                        roundup(details.c / 25)
+                        roundup((details.c || 0) / 25)
                       }
                       onLastClicked={() => {
                         changePage(
@@ -275,29 +332,38 @@ function Conversation(props: { id: number }) {
                       }}
                     />
                   </ReactVisibilitySensor>
-                  {splitarray(
-                    conversation,
-                    index * 25,
-                    (index + 1) * 25 - 1
-                  ).map(
-                    (comment: any) =>
-                      !comment?.removed && (
-                        <Comment
-                          name={users?.[comment?.user].name}
-                          id={comment.id}
-                          op={users?.[comment?.user].name === details.op}
-                          sex={users?.[comment?.user].sex}
-                          date={comment?.createdAt}
-                          tid={props.id}
-                          up={comment?.["U"] | 0}
-                          down={comment?.["D"] | 0}
-                          vote={votes?.[comment.id]}
-                          userid={comment?.user}
-                        >
-                          {DOMPurify.sanitize(comment?.comment)}
-                        </Comment>
-                      )
-                  )}
+                  <ConversationContext.Provider
+                    value={{
+                      story: [story, setStory],
+                      tid: id,
+                      title: details.title,
+                    }}
+                  >
+                    {splitarray(
+                      conversation,
+                      index * 25,
+                      (index + 1) * 25 - 1
+                    ).map(
+                      (comment: any) =>
+                        !comment?.removed &&
+                        (story ? story === comment?.user : 1) && (
+                          <Comment
+                            name={users?.[comment?.user].name}
+                            id={comment.id}
+                            op={users?.[comment?.user].name === details.op}
+                            sex={users?.[comment?.user].sex}
+                            date={comment?.createdAt}
+                            up={comment?.["U"] | 0}
+                            down={comment?.["D"] | 0}
+                            vote={votes?.[comment.id]}
+                            userid={comment?.user}
+                            slink={comment?.slink}
+                          >
+                            {DOMPurify.sanitize(comment?.comment)}
+                          </Comment>
+                        )
+                    )}
+                  </ConversationContext.Provider>
                 </Box>
               ))}
           </Box>
@@ -322,9 +388,22 @@ function Conversation(props: { id: number }) {
               <CircularProgress color="secondary" />
             )}
           </Box>
+          <PageBottom />
         </Paper>
       </div>
     </ShareProvider>
   );
+}
+export function useTid(): number {
+  const { tid } = useContext(ConversationContext);
+  return tid;
+}
+export function useTitle(): string {
+  const { title } = useContext(ConversationContext);
+  return title;
+}
+export function useStory(): [number, React.Dispatch<SetStateAction<number>>] {
+  const { story } = useContext(ConversationContext);
+  return story;
 }
 export default memo(Conversation);