+++ src/components/sidebar.tsx
@@ -24,12 +24,11 @@ import SearchBar from "./searchbar";
 import { useQuery } from "./ContextProvider";
 import { categories, wholepath } from "../lib/common";
 import { useCat, useData, useProfile, useSearch } from "./MenuProvider";
-import MetahkgLogo from "./icon";
-/*
- * The sidebar used by Menu
- * link to metahkg frontpage, search bar, sign in/register/logout,
- * create topic, link to categories, telegram group and source code,
- * at the bottom, if signed in, a link to /profile/self
+import MetahkgLogo from "./logo";
+/**
+ * The sidebar is a
+ * drawer that is opened by clicking on the menu icon on the top left of the
+ * screen. It contains a list of links to different pages
  */
 export default function SideBar() {
   const [open, setOpen] = useState(false);
@@ -75,14 +74,14 @@ export default function SideBar() {
         <Box className="sidebar-box max-width-full" role="presentation">
           <div className="fullwidth">
             <List className="fullwidth">
-              <Link to="/" className="notextdecoration white">
+              <a href="https://war.ukraine.ua/donate/" className="notextdecoration white">
                 <ListItem button onClick={onClick}>
                   <ListItemIcon>
-                    <MetahkgLogo height={24} width={24} svg light />
+                    <MetahkgLogo height={24} width={30} ua />
                   </ListItemIcon>
-                  <ListItemText>Metahkg</ListItemText>
+                  <ListItemText>Donate to Ukraine</ListItemText>
                 </ListItem>
-              </Link>
+              </a>
             </List>
             <div className="ml10 mr10">
               <SearchBar
@@ -182,7 +181,7 @@ export default function SideBar() {
             </div>
           )}
           <p className="ml5">
-            Metahkg build {process.env.REACT_APP_build || "v0.5.3"}
+            Metahkg build {process.env.REACT_APP_build || "v0.5.5"}
           </p>
         </Box>
       </Drawer>