+++ src/pages/create.tsx
@@ -1,5 +1,5 @@
 import "./css/create.css";
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import {
   Alert,
   Box,
@@ -15,6 +15,7 @@ import TextEditor from "../components/texteditor";
 import HCaptcha from "@hcaptcha/react-hcaptcha";
 import axios from "axios";
 import { Navigate, useNavigate } from "react-router";
+import queryString from "query-string";
 import {
   useCat,
   useData,
@@ -24,12 +25,15 @@ import {
 } from "../components/MenuProvider";
 import { useNotification, useWidth } from "../components/ContextProvider";
 import { categories, severity, wholepath } from "../lib/common";
-import MetahkgLogo from "../components/icon";
-declare const hcaptcha: { reset: (e: string) => void }; //the hcaptcha object, defined to use hcaptcha.reload("")
-/*
- * A select list to choose category
- * props.errorHandler: executed if categories cannot to fetched
- * props.changehandler: used as a callback after user changes category selection
+import MetahkgLogo from "../components/logo";
+/* A workaround for the hcaptcha.reset() function not being exported from the hcaptcha library. */
+declare const hcaptcha: { reset: (e: string) => void };
+/**
+ * It takes in a category number and a setter function for the category number, and returns a form
+ * control with a select menu that allows the user to choose a category
+ * @param {number} props.cat The currently choosed category
+ * @param {React.Dispatch<React.SetStateAction<number>>} props.setCat The function to update props.cat
+ * @returns A form control with a select menu.
  */
 function ChooseCat(props: {
   cat: number;
@@ -59,14 +63,12 @@ function ChooseCat(props: {
     </div>
   );
 }
-/*
- * Create component for /create
- * renders a tinymce editor (for content) and a textfield (for title)
- * A captcha must be completed before a user can create a thread
- * The user must be signed in, otherwise he would be redirected to /signin
+/**
+ * Page for creating a new topic
  */
 export default function Create() {
   const navigate = useNavigate();
+  const query = queryString.parse(window.location.search);
   const [menu, setMenu] = useMenu();
   const [width] = useWidth();
   const [profile, setProfile] = useProfile();
@@ -83,6 +85,37 @@ export default function Create() {
     severity: "info",
     text: "",
   });
+  const quote = {
+    id: Number(String(query.quote).split(".")[0]),
+    cid: Number(String(query.quote).split(".")[1]),
+  };
+  const [inittext, setInittext] = useState("");
+  /**
+   * It sends data to the /api/create route.
+   */
+  useEffect(() => {
+    if (localStorage.user && quote.id && quote.cid) {
+      setNotification({ open: true, text: "Fetching comment..." });
+      axios
+        .get(
+          `/api/thread/${quote.id}?type=2&start=${quote.cid}&end=${quote.cid}`
+        )
+        .then((res) => {
+          setInittext(
+            `<blockquote style="color: #aca9a9; border-left: 2px solid #aca9a9; margin-left: 0"><div style="margin-left: 15px">${res.data?.[0]?.comment}</div></blockquote><p></p>`
+          );
+          setTimeout(() => {
+            setNotification({ open: false, text: "" });
+          }, 1000);
+        })
+        .catch(() => {
+          setNotification({
+            open: true,
+            text: "Unable to fetch comment. This comment would not be a quote.",
+          });
+        });
+    }
+  }, [quote.cid, quote.id, setNotification]);
   function create() {
     //send data to /api/create
     setAlert({ severity: "info", text: "Creating topic..." });
@@ -116,14 +149,13 @@ export default function Create() {
   }
   document.title = "Create topic | Metahkg";
   menu && setMenu(false);
-  if (!localStorage.user) {
+  if (!localStorage.user)
     return (
       <Navigate
         to={`/signin?continue=true&returnto=${encodeURIComponent(wholepath())}`}
         replace
       />
     );
-  }
   const small = width * 0.8 - 40 <= 450;
   return (
     <Box
@@ -163,7 +195,7 @@ export default function Create() {
             changehandler={(v, e: any) => {
               setIcomment(e.getContent());
             }}
-            text=""
+            text={inittext}
           />
           <div className="mt20">
             <ChooseCat cat={catchoosed} setCat={setCatchoosed} />