+++ src/pages/notfound.tsx
@@ -1,8 +1,11 @@
 import "./css/notfound.css";
 import React from "react";
 import { Box } from "@mui/material";
-import MetahkgLogo from "../components/icon";
+import MetahkgLogo from "../components/logo";
 import { useMenu } from "../components/MenuProvider";
+/**
+ * 404 page 
+ */
 export default function NotFound() {
   document.title = "404 Not Found | Metahkg";
   const [menu, setMenu] = useMenu();